requirements.txt
GitPython==3.1.24
requests==2.22.0
